Use 8-bit instead of 16-bit loop indices where they will suffice

+++ right/src/config_parser/parse_config.c
@@ -27,7 +27,7 @@ parser_error_t ParseConfig(config_buffer_t *buffer)
     uint16_t keymapCount;
 
     (void)dataModelVersion;
-    for (uint16_t moduleConfigurationIdx = 0; moduleConfigurationIdx < moduleConfigurationCount; moduleConfigurationIdx++) {
+    for (uint8_t moduleConfigurationIdx = 0; moduleConfigurationIdx < moduleConfigurationCount; moduleConfigurationIdx++) {
         errorCode = parseModuleConfiguration(buffer);
         if (errorCode != ParserError_Success) {
             return errorCode;
@@ -41,7 +41,7 @@ parser_error_t ParseConfig(config_buffer_t *buffer)
         }
     }
     keymapCount = readCompactLength(buffer);
-    for (uint16_t keymapIdx = 0; keymapIdx < keymapCount; keymapIdx++) {
+    for (uint8_t keymapIdx = 0; keymapIdx < keymapCount; keymapIdx++) {
         errorCode = ParseKeymap(buffer, keymapIdx, keymapCount);
         if (errorCode != ParserError_Success) {
             return errorCode;
